Fix debouncing on normal input
